+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### VNEXT
 
 * Made `resolvers` parameter optional for `mergeSchemas` [Issue #461](https://github.com/apollographql/graphql-tools/issues/461) [PR #462] (https://github.com/apollographql/graphql-tools/pull/462)
+* Added basic subscription support for local schemas [Issue #420](https://github.com/apollographql/graphql-tools/issues/420) [PR #463] (https://github.com/apollographql/graphql-tools/pull/463)
 
 ### v2.7.0